chore: refactor ssh commands
