--- auth_api/commands/ssh_ro_sftp.py
@@ -1,1 +1,0 @@
-ssh_no_mfa.py
