--- auth_api/commands/ssh_tre_sftp.py
@@ -1,1 +1,0 @@
-ssh_no_mfa.py
